service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv  # To load .env file implicitly with pydantic-settings
loguru
requests